requirements.txt
google-generativeai
pandas
pytest
streamlit
